cat vector types by kind
Rearrange aliases to bottom of files
