+++ crates/core_simd/src/vector.rs
@@ -1,27 +1,3 @@
-mod vectors_f64;
-mod vectors_i128;
-mod vectors_i16;
-mod vectors_i32;
-mod vectors_i64;
-mod vectors_i8;
-mod vectors_u128;
-mod vectors_u16;
-mod vectors_u32;
-mod vectors_u64;
-mod vectors_u8;
-
-pub use vectors_f64::*;
-pub use vectors_i128::*;
-pub use vectors_i16::*;
-pub use vectors_i32::*;
-pub use vectors_i64::*;
-pub use vectors_i8::*;
-pub use vectors_u128::*;
-pub use vectors_u16::*;
-pub use vectors_u32::*;
-pub use vectors_u64::*;
-pub use vectors_u8::*;
-
 mod float;
 mod int;
 mod uint;
